agentServer/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autogen-agentchat==0.7.3
autogen-core==0.7.3
autogen-ext==0.7.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
google-auth==2.40.3
google-genai==1.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mcp==1.13.0
openai==1.100.1
opentelemetry-api==1.36.0
pillow==11.3.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rpds-py==0.27.0
rsa==4.9.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
zipp==3.23.0
